+++ src/merchant.rs
@@ -2,5 +2,5 @@ pub use crate::cli::{merchant as cli, merchant::Cli};
 pub use crate::config::{merchant as config, merchant::Config};
 pub use crate::database::merchant as database;
 pub use crate::defaults::merchant as defaults;
-pub use crate::transport::server::{self as server, Chan, Server};
 pub use crate::zkchannels::merchant as zkchannels;
+pub use transport::server::{self as server, Chan, Server};
